Rename and add @Repository tag
JIRA: GLT-1290
